// src/enum/git-commands.enum.ts
export enum GitCommandNamesEnum {
  add = 'add',
  addForce = 'addForce',
  assumeUnchanged = 'assumeUnchanged',
  noAssumeUnchanged = 'noAssumeUnchanged',
  checkInitialized = 'checkInitialized',
  status = 'status',
  branch = 'branch',
  checkout = 'checkout',
  clone = 'clone',
  commit = 'commit',
  config = 'config',
  diff = 'diff',
  fetch = 'fetch',
  init = 'init',
  log = 'log',
  merge = 'merge',
  pull = 'pull',
  push = 'push',
  rebase = 'rebase',
  remote = 'remote',
  reset = 'reset',
  rm = 'rm',
  show = 'show',
  tag = 'tag',
  stash = 'stash',
  submodule = 'submodule',
  worktree = 'worktree',
}

export enum GitCommandsEnum {
  add = 'add',
  addForce = 'add --force',
  assumeUnchanged = 'update-index --assume-unchanged',
  noAssumeUnchanged = 'update-index --no-assume-unchanged',
  checkInitialized = 'rev-parse --is-inside-work-tree',
  status = 'status --short',
  branch = 'branch',
  checkout = 'checkout',
  clone = 'clone',
  commit = 'commit',
  config = 'config',
  diff = 'diff',
  fetch = 'fetch',
  init = 'init',
  log = 'log',
  merge = 'merge',
  pull = 'pull',
  push = 'push',
  rebase = 'rebase',
  remote = 'remote',
  reset = 'reset',
  rm = 'rm',
  show = 'show',
  tag = 'tag',
  stash = 'stash',
  submodule = 'submodule',
  worktree = 'worktree',
}